requirements.txt
pyTelegramBotAPI==4.14.0
flask
